+++ app/zpk/http/zpk.go
@@ -941,3 +941,15 @@ func (self Zpk) LocalZpkUrl(http *gin.Context) {
 		"oauthToken": oauthToken,
 	})
 }
+
+// DomainParse 返回主集群的域名解析配置，子集群安装应用时也应使用此配置提示用户。
+func (self Zpk) DomainParse(http *gin.Context) {
+	client := k8s.NewK8sClient()
+
+	config, err := client.GetConfigCRD(http.Request.Context(), k8s.DomainParseConfigGVR, k8s.DomainParseConfigName)
+	if err != nil {
+		self.JsonResponseWithServerError(http, err)
+		return
+	}
+	self.JsonResponseWithoutError(http, config)
+}